Use gxps_value_get_double_positive() to parse page sizes
And report an error if Width or Height are missing or fail to parse
since they are required attributes of FixedPage. Even though all XPS
documents we have for testing use integer values for page sizes, the
spec says with/height are real numbers greater or equal than 1.
This commit breaks the API because gxps_document_get_page_size() and
gxps_page_get_size() return gdouble instead of guint values now.

+++ libgxps/gxps-document.h
@@ -68,8 +68,8 @@ GXPSPage              *gxps_document_get_page            (GXPSDocument *doc,
 							  GError      **error);
 gboolean               gxps_document_get_page_size       (GXPSDocument *doc,
 							  guint         n_page,
-							  guint        *width,
-							  guint        *height);
+							  gdouble      *width,
+							  gdouble      *height);
 gint                   gxps_document_get_page_for_anchor (GXPSDocument *doc,
 							  const gchar  *anchor);
 GXPSDocumentStructure *gxps_document_get_structure       (GXPSDocument *doc);
